Studio Code: move the design catalog out of the skills registry into its own module
