+++ apps/cli/ai/tools/skill.ts
@@ -1,5 +1,6 @@
 import { Type } from 'typebox';
-import { findSkill, loadSkills, renderSkillBody } from 'cli/ai/skills';
+import { renderDesignCatalogIndex } from 'cli/ai/design-catalog';
+import { findSkill, loadSkills } from 'cli/ai/skills';
 import { defineTool } from './define-tool';
 import type { AgentTool } from '@earendil-works/pi-agent-core';
 import type { TSchema } from 'typebox';
@@ -25,7 +26,7 @@ export function createSkillTool(): AgentTool< TSchema > | null {
 				throw new Error( `Unknown skill: ${ args.name }` );
 			}
 			return {
-				content: [ { type: 'text' as const, text: renderSkillBody( skill ) } ],
+				content: [ { type: 'text' as const, text: renderDesignCatalogIndex( skill.body ) } ],
 			};
 		}
 	);
